requirements.txt
PyGithub==1.45
StackAPI==0.1.12